refactor(metamask): change folder struct
